+++ frontend/src/__tests__/snapshots/SignUpModal.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import { render } from '@testing-library/react';
-import SignUpModal from '../../components/SignUpModal';
+import SignUpModal from '../../components/FamilySignUpModal';
 
 function onClose() {
     // a void function
