+++ frontend/src/components/AccountPageTabs.tsx
@@ -20,6 +20,7 @@ import { Bike } from '../../../src/typedefs/bike';
 
 const memberFamily: Member[] = [
     {
+        membershipId: 1,
         memberId: 1,
         membershipAdmin: 'true',
         active: true,
@@ -40,6 +41,7 @@ const memberFamily: Member[] = [
         lastModifiedBy: '',
     },
     {
+        membershipId: 1,
         memberId: 2,
         membershipAdmin: 'true',
         active: true,
