+++ frontend/src/components/GeneralInfo.tsx
@@ -30,6 +30,7 @@ async function handlePatchMemberContactInfo(
     // }
 
     const updatedMember: Member = {
+        membershipId: 1,
         memberId: 1,
         membershipAdmin: 'true',
         active: true,
