+++ frontend/src/components/MemberSummaryModal.tsx
@@ -72,6 +72,7 @@ async function handlePatchMemberContactInfo(
     // }
 
     const updatedMember: Member = {
+        membershipId: 1,
         memberId: 1,
         membershipAdmin: 'true',
         active: true,
